BsMD-package {BsMD}	R Documentation

Bayes screening and model discrimination follow-up designs

Description

Bayes screening and model discrimination follow-up designs

Details

Package: 	 BsMD

Type: 	 Package

Version: 	 2013.0718

Date: 	 2013-07-18

License: 	 GPL version 2 or later

The packages allows you to perform the calculations and analyses described in Mayer, Stainberg and Box paper in Technometrics, 1996.

Author(s)

Author: Ernesto Barrios based on Daniel Meyer's code. Maintainer: Ernesto Barrios <ebarrios@itam.mx>

References

Box and Mayer, 1986; Box and Mayer, 1993; Mayer, Steinberg and Box, 1996.

Examples

    data(BM86.data)

[Package BsMD version 2013.0718-1 Index]