summary.BsProb {BsMD}	R Documentation

Summary of Posterior Probabilities from Bayesian Screening

Description

Reduced printing method for class BsProb lists. Prints posterior probabilities of factors and models from Bayesian screening procedure.

Usage

    ## S3 method for class 'BsProb'
summary(object, nMod = 10, digits = 3, ...)

Arguments

object
list. BsProb class list. Output list of BsProb function.

nMod
integer. Number of the top ranked models to print.

digits
integer. Significant digits to use.

...
additional arguments passed to summary generic function.

Value

The function prints out the marginal factors and models posterior probabilities. Returns invisible list with the components:

calc
Numeric vector with basic calculation information.

probabilities
Data frame with the marginal posterior factor probabilities.

models
Data frame with the models posterior probabilities.

Author(s)

Ernesto Barrios.

References

Box, G. E. P and R. D. Meyer (1986). "An Analysis for Unreplicated Fractional Factorials". Technometrics. Vol. 28. No. 1. pp. 11–18.

Box, G. E. P and R. D. Meyer (1993). "Finding the Active Factors in Fractionated Screening Experiments". Journal of Quality Technology. Vol. 25. No. 2. pp. 94–105.

See Also

BsProb, print.BsProb, plot.BsProb.

Examples

library(BsMD)
data(BM86.data,package="BsMD")
X <- as.matrix(BM86.data[,1:15])
y <- BM86.data["y1"]
# Using prior probability of p = 0.20, and k = 10 (gamma = 2.49)
drillAdvance.BsProb <- BsProb(X = X, y = y, blk = 0, mFac = 15, mInt = 1,
            p = 0.20, g = 2.49, ng = 1, nMod = 10)
plot(drillAdvance.BsProb)
summary(drillAdvance.BsProb)

# Using prior probability of p = 0.20, and a 5 <= k <= 15 (1.22 <= gamma <= 3.74)
drillAdvance.BsProbG <- BsProb(X = X, y = y, blk = 0, mFac = 15, mInt = 1,
            p = 0.25, g = c(1.22, 3.74), ng = 3, nMod = 10)
plot(drillAdvance.BsProbG)
summary(drillAdvance.BsProbG)

[Package BsMD version 2013.0718-1 Index]